requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
langchain-huggingface
unstructured
faiss-cpu
python-dotenv
